refactor credits endpoint
