+++ src/components/RecipeForm.jsx
@@ -95,7 +95,7 @@ function RecipeForm({
       return;
     }
     try {
-      const res = await fetch('/api/get-ia-credits', {
+      const res = await fetch('/api/credits', {
         headers: {
           Authorization: `Bearer ${session.access_token}`,
           apikey: import.meta.env.VITE_SUPABASE_ANON_KEY,
