+++ src/components/account/IACredits.jsx
@@ -15,7 +15,7 @@ export default function IACredits({ session }) {
   const fetchCredits = async () => {
     if (!session?.access_token) return;
     try {
-      const res = await fetch('/api/get-ia-credits', {
+      const res = await fetch('/api/credits', {
         headers: {
           Authorization: `Bearer ${session.access_token}`,
           apikey: import.meta.env.VITE_SUPABASE_ANON_KEY,
@@ -74,7 +74,7 @@ export default function IACredits({ session }) {
     setLoading(type);
     try {
       const priceId = type === 'text' ? TEXT_PRICE_ID : IMAGE_PRICE_ID;
-      const res = await fetch('/api/purchase-credits', {
+      const res = await fetch('/api/credits', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
